requirements.txt
Brotli==1.0.9
click==7.1.2
dash==1.20.0
dash-core-components==1.16.0
dash-daq==0.5.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
Flask==1.1.2
Flask-Compress==1.9.0
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.24
plotly==4.14.3
protobuf==3.16.0
PyMySQL==1.0.2
retrying==1.3.3
six==1.16.0
Werkzeug==1.0.1
